add main_page
